+++ packages/core/src/components/lf-canvas/elements.canvas.tsx
@@ -1,4 +1,8 @@
-import { LfCanvasAdapter, LfCanvasAdapterJsx } from "@lf-widgets/foundations";
+import {
+  LfCanvasAdapter,
+  LfCanvasAdapterJsx,
+  LfImageEventPayload,
+} from "@lf-widgets/foundations";
 import { h } from "@stencil/core";
 
 export const prepCanvasJsx = (
@@ -32,10 +36,20 @@ export const prepCanvasJsx = (
 
     //#region Image
     image: () => {
-      const { controller, elements } = getAdapter();
-      const { refs } = elements;
+      const { controller, elements, handlers } = getAdapter();
       const { blocks, compInstance, cyAttributes, manager, parts } =
         controller.get;
+      const { refs } = elements;
+      const { onLoad } = handlers.image;
+      const onImageEvent = (event: CustomEvent<LfImageEventPayload>) => {
+        const { eventType } = event.detail;
+
+        switch (eventType) {
+          case "load":
+            void onLoad(event);
+            break;
+        }
+      };
       const { assignRef, sanitizeProps, theme } = manager;
       const { bemClass } = theme;
 
@@ -44,6 +58,7 @@ export const prepCanvasJsx = (
           {...sanitizeProps(compInstance.lfImageProps, "LfImage")}
           class={bemClass(blocks.canvas._, blocks.canvas.image)}
           data-cy={cyAttributes.image}
+          onLf-image-event={onImageEvent}
           part={parts.image}
           ref={assignRef(refs, "image")}
         ></lf-image>