+++ packages/core/src/components/lf-compare/helpers.utils.ts
@@ -15,11 +15,13 @@ export const prepTreeDataset = (
   for (let index = 0; index < shapes.length; index++) {
     const shape = shapes[index];
     const strIndex = String(index).valueOf();
+    const hasValue =
+      typeof shape.value === "string" && shape.value.trim().length > 0;
 
     dataset.nodes.push({
       icon: currentShape === shape && currentIcon,
       id: strIndex,
-      value: `#${strIndex}`,
+      value: hasValue ? shape.value : `#${strIndex}`,
     });
   }