+++ packages/framework/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0-rc.2](https://github.com/lucafoscili/lf-widgets/compare/0.5.0-rc.1...0.5.0-rc.2) (2025-09-30)
+
+**Note:** Version bump only for package @lf-widgets/framework
+
+# [0.5.0-rc.1](https://github.com/lucafoscili/lf-widgets/compare/0.5.0-rc.0...0.5.0-rc.1) (2025-09-30)
+
+**Note:** Version bump only for package @lf-widgets/framework
+
+# [0.5.0-rc.0](https://github.com/lucafoscili/lf-widgets/compare/0.4.2-rc.1...0.5.0-rc.0) (2025-09-30)
+
+**Note:** Version bump only for package @lf-widgets/framework
+
+## [0.4.2-rc.1](https://github.com/lucafoscili/lf-widgets/compare/0.4.2-rc.0...0.4.2-rc.1) (2025-09-29)
+
+**Note:** Version bump only for package @lf-widgets/framework
+
+## [0.4.2-rc.0](https://github.com/lucafoscili/lf-widgets/compare/0.4.1...0.4.2-rc.0) (2025-09-29)
+
+**Note:** Version bump only for package @lf-widgets/framework
+
 ## [0.4.1](https://github.com/lucafoscili/lf-widgets/compare/0.4.1-rc.0...0.4.1) (2025-09-23)
 
 **Note:** Version bump only for package @lf-widgets/framework